Improve PollModule.cs (#39)
* Update PollModule.cs

Removed the StringBuilder in poll.
Added the command endpoll. This command ends the poll by editing the original poll. It shows the votes of every option.

* Updated requested changes on #39

Some things...

* Added check for endpoll

If you try to endpoll messages which don't match the poll rules (so they aren't polls), you will get an error :-)

* Readd StringBuilder + Extract embed to var

Because of performance reason the StringBuilder got back, and also the embed in the check method is now extracted.

* Split if in check method

The if in the check method is now spiltted to multiple ifs.

* Add space + Take out fields

Because of performance reason, the fields got limited to 2.

* FirstOrDefault -> First

* Changed method to regular index 0

.First() to [0]